test ok

+++ tp5/LOG4420_TP5/client/src/app/shopping-cart/shopping-cart.component.ts
@@ -32,17 +32,24 @@ export class ShoppingCartComponent implements OnInit {
     this.updateTotal();
   }
   clearRow(item){
-    let idx = this.sc.indexOf(item);
-    if(idx !== -1){
-      this.sc.splice(idx,1);
+    var confirmBox = confirm("Voulez-vous supprimer le produit du panier ?");
+    if (confirmBox == true) {
+      let idx = this.sc.indexOf(item);
+      if(idx !== -1){
+        this.sc.splice(idx,1);
+      }
+      this.apiService.deleteDataWithPromiseShoppingCart(item.product.id);
+      this.updateTotal();
     }
-    this.apiService.deleteDataWithPromiseShoppingCart(item.product.id);
-    this.updateTotal();
+ 
   }
   emptySC(){
-    this.sc = [];
-    this.apiService.deleteAllDataWithPromiseShoppingCart();
-    this.updateTotal();
+    var confirmBox = confirm("Voulez-vous supprimer tous les produits du panier ?");
+    if (confirmBox == true) {
+      this.sc = [];
+      this.apiService.deleteAllDataWithPromiseShoppingCart();
+      this.updateTotal();
+    }
   }
   updateTotal(){
     this.total = 0;
